Refactor: move setup_requests because of circular imports
Move setup_requests to a place less likely to cause circular imports.
