+++ openlibrary/plugins/upstream/utils.py
@@ -56,6 +56,8 @@ if TYPE_CHECKING:
 STRIP_CHARS = ",'\" "
 REPLACE_CHARS = "]["
 
+logger = logging.getLogger('openlibrary')
+
 
 class LanguageMultipleMatchError(Exception):
     """Exception raised when more than one possible language match is found."""
@@ -1613,6 +1615,28 @@ def get_location_and_publisher(loc_pub: str) -> tuple[list[str], list[str]]:
 
     # Fall back to making the input a list returning that and an empty location.
     return ([], [loc_pub.strip(STRIP_CHARS)])
+
+
+def setup_requests():
+    logger.info("Setting up requests")
+
+    logger.info("Setting up proxy")
+    if config.get("http_proxy", ""):
+        os.environ['HTTP_PROXY'] = os.environ['http_proxy'] = config.get('http_proxy')
+        os.environ['HTTPS_PROXY'] = os.environ['https_proxy'] = config.get('http_proxy')
+        logger.info('Proxy environment variables are set')
+    else:
+        logger.info("No proxy configuration found")
+
+    logger.info("Setting up proxy bypass")
+    if config.get("no_proxy_addresses", []):
+        no_proxy = ",".join(config.get("no_proxy_addresses"))
+        os.environ['NO_PROXY'] = os.environ['no_proxy'] = no_proxy
+        logger.info('Proxy bypass environment variables are set')
+    else:
+        logger.info("No proxy bypass configuration found")
+
+    logger.info("Requests set up")
 
 
 def setup() -> None:
