+++ scripts/affiliate_server.py
@@ -57,7 +57,7 @@ from openlibrary.config import load_config as openlibrary_load_config
 from openlibrary.core import cache, stats
 from openlibrary.core.imports import Batch, ImportItem
 from openlibrary.core.vendors import AmazonAPI, clean_amazon_metadata_for_load
-from openlibrary.plugins.openlibrary.code import setup_requests
+from openlibrary.plugins.upstream.utils import setup_requests
 from openlibrary.utils.dateutil import WEEK_SECS
 from openlibrary.utils.isbn import (
     isbn_10_to_isbn_13,
