+++ scripts/partner_batch_imports.py
@@ -22,7 +22,7 @@ import requests
 from infogami import config  # noqa: F401 side effects may be needed
 from openlibrary.config import load_config
 from openlibrary.core.imports import Batch
-from openlibrary.plugins.openlibrary.code import setup_requests
+from openlibrary.plugins.upstream.utils import setup_requests
 from scripts.solr_builder.solr_builder.fn_to_cli import FnToCLI
 
 logger = logging.getLogger("openlibrary.importer.bwb")
